requirements.txt
python-binance==1.0.19
python-telegram-bot==13.15 